logging: switch to klog/v2
If klog and klog/v2 are mixed in the same binary and the binary only
initializes the klog command line flags, then output via klog/v2
doesn't use the log level of the binary.

Most of the code, with the exception of the
sig-storage-lib-external-provisioner, is now using klog/v2, so it
makes more sense to switch to that and ensure that output via that
package gets the right log level.

Long-term klog should get removed entirely.